Fix function getMethodsActivedsMelhorEnvio()
This function check if exist a Melhor Envio shipping method in all shipping zones, but if the store is set to use a Melhor Envio shipping method in all situations (without a zone), the plugin show a alert message that don't exist a shipping method configured. This fix change the function to check if exist a Melhor Envio method active, independent if it is in a zone or not.